requirements.txt
asgiref==3.5.1
autopep8==1.6.0
bidict==0.22.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
distlib==0.3.4
Django==4.0.4
dnspython==2.2.1
eventlet==0.33.1
filelock==3.6.0
Flask==2.1.2
Flask-SocketIO==5.1.2
greenlet==1.1.2
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
Pillow==9.1.0
platformdirs==2.5.2
pycodestyle==2.8.0
PySocks==1.7.1
python-dotenv==0.20.0
python-engineio==4.3.2
python-socketio==5.6.0
requests==2.27.1
requests-futures==1.0.0
six==1.16.0
sqlparse==0.4.2
stem==1.8.0
toml==0.10.2
torrequest==0.1.0
urllib3==1.26.9
virtualenv==20.14.1
Werkzeug==2.1.2
whitenoise==6.1.0

djangorestframework~=3.13.1
pip~=22.0.4
setuptools~=58.1.0